wrapper: Fix size type
(cherry picked from commit f8e68b92b83c67043ee81c81ffbe817488420d1f)

+++ src/wrapper.c
@@ -148,7 +148,8 @@ struct ssh_crypto_struct *crypto_new(void) {
 
 void crypto_free(struct ssh_crypto_struct *crypto)
 {
-    int i;
+    size_t i;
+
     if (crypto == NULL) {
         return;
     }
